Revert "patient home screen"
